+++ blocks/canvas/ew-editor-wysiwyg/utils/handlers.js
@@ -6,7 +6,7 @@ import {
   NX_QUICK_EDIT_IFRAME_SELECTION_META,
   NX_QUICK_EDIT_CLEAR_IFRAME_SELECTION_ORIGIN_META,
 } from '../../editor-utils/selection-toolbar.js';
-import { editorSelectChange } from '../../editor-utils/document.js';
+import { editorSelectChange } from '../../editor-utils/editor-utils.js';
 import { getActiveBlockFlatIndex } from './blocks.js';
 
 export function handleCursorMove({ cursorOffset, textCursorOffset }, ctx) {
